refactor: restructure provider-inventory entities to better reflect layers
