+++ apps/provider-inventory/src/domain/resource-pair/resource-pair.spec.ts
