+++ apps/provider-inventory/src/lib/open-api-hono-handler/open-api-hono-handler.ts
