+++ apps/provider-inventory/src/mappers/gpu-attribute-parser/gpu-attribute-parser.spec.ts
