+++ apps/provider-inventory/src/mappers/gpu-attribute-parser/gpu-attribute-parser.ts
@@ -1,4 +1,4 @@
-import type { ResourceAttribute } from "../../types/inventory.types";
+import type { ResourceAttribute } from "../../types/inventory";
 
 export interface ParsedGPUAttributes {
   vendor: string;
