+++ apps/provider-inventory/src/mappers/storage-attribute-parser/storage-attribute-parser.spec.ts
