--- apps/provider-inventory/src/types/provider.ts
@@ -1,7 +1,0 @@
-import type { ClusterState } from "./inventory.types";
-
-export interface ProviderWithClusterState {
-  owner: string;
-  hostUri: string;
-  cluster?: ClusterState;
-}
